responsive (x-axis parallax) slider complete
fixed the `overflow-x:hidden` compatibility issue caused by slider for Chrome by applying that declaration to even the `html` element (see `index--base.css`).

But this caused a problem:
the scroll for the menu overlay wasn't  disabling, so I changed the nav.js a bit: changed the selector for `body` to `html` and went to `nav.css` file and changed the `body.noscroll` to `html.noscroll`. Scrolling disabled on Chrome and FireFox now.